Refine GW2 login
Revert a change in the OAuth library while adding the possibility to add additional headers to single calls. Also log full request details if logging is on.

Also use the new "additional header" parameter in the GuildWars provider

I didn't add an image to the guild wars 2 button, because I don't know how...